hide lockup dropdown in stake delegation when staking is disabled
